Removed campaign and event from NGReportFactory.
Both 'campaign' and 'event' are not required fields of an NGReport and
should not be populated automatically or unexpected things may
happen. For example creating an NGReport using the factory, would in
fact create two reports, the second of which was triggered from the
event created by the Event subfactory.

Similarity campaigns should not be auto-populated, since we expect
campaign field to be not null when activity is set to the special
'Participated in a campaign' string and not just a random Activity.