[fix bug 971632] Do not explicitly define bug component in event form.
When attaching a bug number to events, as Budget or Swag request, the
form relates it to a Bug objects in remozilla.

Because users often mix the Budget with the Swag field -and the other
way around-, the event form tries to find a Bug in remozilla using the
correct id and the wrong category, which will create a new Bug object
with these bogus data.

The fix is to lookup the bug using on the bug id, since Bugzilla
guaranties that the id is unique.