[fix bug 787791] Stop making API calls when full list gets loaded.
Tastypie upgrade (8757ea5) changed the API behavior when it comes to
'limit' meta data returned to API call. It's now limited to 1000, while
previously respected any limit given. We based parts of our logic on the
limit returned and that caused Events and Reps to be loaded multiple
times by consecutive API calls.

We fix that by watching the meta.next variable which always tells the
truth about the existence of more data, independently of the value of
the limit.